% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getPackageDate}
\alias{getPackageDate}
\alias{getPackageDate,character-method}
\alias{getPackageDate,missing-method}
\title{Get Package Date}
\usage{
getPackageDate(from = ".", ...)

\S4method{getPackageDate}{character}(from = ".", ...)

\S4method{getPackageDate}{missing}(from = ".", ...)
}
\arguments{
\item{from}{\strong{Signature argument}.
Object containing source information.}

\item{...}{Further arguments passed to subsequente functions:
\code{\link[rapp.core.description]{getPackageDescription}}.}

\item{from}{\code{\link{character}}.}

\item{from}{\code{\link{missing}}.}
}
\value{
\code{character}.

See signature \code{character}.
}
\description{
Get package date.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: 

\item \code{missing}: 
}}
\examples{
##------------------------------------------------------------------------------
## From package project DESCRIPTION file //
##------------------------------------------------------------------------------

## This expects the existence of a DESCRIPTION file in your current 
## working directory (else it will fail!):
getPackageDate()  

## Explicit specification of the DESCRIPTION file's location: 
getPackageDate(from = "DESCRIPTION")  

##------------------------------------------------------------------------------
## From package DESCRIPTION file //
##------------------------------------------------------------------------------

require("devtools")
getPackageDate(from = "devtools")  
## Note that the field name has been inferred as 'devtools' does not use 
## the field 'Date' but 'Date/Publication'

## Error handling //
getPackageDate(from = tempdir()) 
expect_equal(res, NA)  
    

}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp.core.description}
}

