% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getPackageDependentDependencies}
\alias{getPackageDependentDependencies}
\alias{getPackageDependentDependencies,character-method}
\alias{getPackageDependentDependencies,missing-method}
\title{Get Package Dependent Dependencies}
\usage{
getPackageDependentDependencies(from = ".", type = c("as_is", "names_only",
  "detailed"), ...)

\S4method{getPackageDependentDependencies}{character}(from = ".",
  type = c("as_is", "names_only", "detailed"), ...)

\S4method{getPackageDependentDependencies}{missing}(from = ".",
  type = c("as_is", "names_only", "detailed"), ...)
}
\arguments{
\item{from}{\strong{Signature argument}.
Object containing source information.}

\item{type}{\code{character}.
Either one of \code{c("as_is", "names_only", "detailed")}.}

\item{...}{Further arguments passed to subsequente functions:
\code{\link[rapp.core.description]{getPackageDescription}}.}

\item{from}{\code{\link{character}}.}

\item{from}{\code{\link{missing}}.}
}
\value{
\code{character}.

See signature \code{character}.
}
\description{
Get package dependencies as stated via the \code{Depends} section.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: 

\item \code{missing}: 
}}
\examples{

}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp.core.description}
}

