% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getPackageDescription}
\alias{getPackageDescription}
\alias{getPackageDescription,character,character-method}
\alias{getPackageDescription,character,missing-method}
\alias{getPackageDescription,missing,character-method}
\alias{getPackageDescription,missing,missing-method}
\title{Get Package Description}
\usage{
getPackageDescription(from = ".", fields = character(), drop = TRUE,
  encoding = "", strict = FALSE, ...)

\S4method{getPackageDescription}{character,character}(from = ".",
  fields = character(), drop = TRUE, encoding = "", strict = FALSE, ...)

\S4method{getPackageDescription}{missing,character}(from = ".",
  fields = character(), drop = TRUE, encoding = "", strict = FALSE, ...)

\S4method{getPackageDescription}{missing,missing}(from = ".",
  fields = character(), drop = TRUE, encoding = "", strict = FALSE, ...)

\S4method{getPackageDescription}{character,missing}(from = ".",
  fields = character(), drop = TRUE, encoding = "", strict = FALSE, ...)
}
\arguments{
\item{from}{\strong{Signature argument}.
Object containing source information.}

\item{fields}{\code{Signature argument}.
Object containing field name information.}

\item{drop}{\code{logical}.
If \code{TRUE} and the length of \code{fields} is 1, then a single
character string with the value of the respective field is returned
instead of an object of class \code{packageDescription}.}

\item{encoding}{\code{character}.
If there is an Encoding field, to what encoding should re-encoding be
attempted? If NA, no re-encoding. The other values are as used by
\code{\link{iconv}}, so the default \code{""} indicates the encoding of
the current locale.}

\item{strict}{\code{logical}.
\code{TRUE}: certain constellations trigger an error. \code{FALSE} (default): these
conditions are only reflected by a return value of \code{NA} and
possibly a corresponding warning.}

\item{...}{Further arguments passed to subsequente functions:
\code{\link{[utils]packageDescription}}.}

\item{from}{\code{\link{character}}.}

\item{fields}{\code{\link{character}}.}

\item{from}{\code{\link{missing}}.}

\item{fields}{\code{\link{character}}.}

\item{from}{\code{\link{missing}}.}

\item{fields}{\code{\link{missing}}.}

\item{from}{\code{\link{character}}.}

\item{fields}{\code{\link{missing}}.}
}
\value{
\code{character}.

See signature \code{character,character}.

See signature \code{character,character}.

See signature \code{character,character}.
}
\description{
Get package description.
}
\section{Methods (by class)}{
\itemize{
\item \code{from = character,fields = character}: 

\item \code{from = missing,fields = character}: 

\item \code{from = missing,fields = missing}: 

\item \code{from = character,fields = missing}: 
}}
\examples{

}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp.core.repos}
}

