% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getPackageDescriptionFromFile}
\alias{getPackageDescriptionFromFile}
\alias{getPackageDescriptionFromFile,character,character-method}
\alias{getPackageDescriptionFromFile,character,missing-method}
\alias{getPackageDescriptionFromFile,missing,character-method}
\alias{getPackageDescriptionFromFile,missing,missing-method}
\title{Get Package Description From File}
\usage{
getPackageDescriptionFromFile(from = "DESCRIPTION", fields = character(),
  drop = TRUE, encoding = "", strict = FALSE, validate = TRUE)


  \S4method{getPackageDescriptionFromFile}{character,character}(from = "DESCRIPTION",
  fields = character(), drop = TRUE, encoding = "", strict = FALSE,
  validate = TRUE)


  \S4method{getPackageDescriptionFromFile}{missing,character}(from = "DESCRIPTION",
  fields = character(), drop = TRUE, encoding = "", strict = FALSE,
  validate = TRUE)


  \S4method{getPackageDescriptionFromFile}{missing,missing}(from = "DESCRIPTION",
  fields = character(), drop = TRUE, encoding = "", strict = FALSE,
  validate = TRUE)


  \S4method{getPackageDescriptionFromFile}{character,missing}(from = "DESCRIPTION",
  fields = character(), drop = TRUE, encoding = "", strict = FALSE,
  validate = TRUE)
}
\arguments{
\item{from}{\strong{Signature argument}.
Object containing file path information.}

\item{fields}{\code{Signature argument}.
Object containing field name information.}

\item{drop}{\code{logical}.
If \code{TRUE} and the length of \code{fields} is 1, then a single
character string with the value of the respective field is returned
instead of an object of class \code{packageDescription}.}

\item{encoding}{\code{character}.
If there is an Encoding field, to what encoding should re-encoding be
attempted? If NA, no re-encoding. The other values are as used by
\code{\link{iconv}}, so the default \code{""} indicates the encoding of
the current locale.}

\item{strict}{\code{logical}.
\code{TRUE}: certain constellations trigger an error. \code{FALSE} (default): these
conditions are only reflected by a return value of \code{NA} and
possibly a corresponding warning.}

\item{validate}{\code{logical}.
Perform validation checks (\code{TRUE}, default) or not (\code{FALSE}).}

\item{from}{\code{\link{character}}.}

\item{fields}{\code{\link{character}}.}

\item{from}{\code{\link{missing}}.}

\item{fields}{\code{\link{character}}.}

\item{from}{\code{\link{missing}}.}

\item{fields}{\code{\link{missing}}.}

\item{from}{\code{\link{character}}.}

\item{fields}{\code{\link{missing}}.}
}
\value{
\code{character}.

See signature \code{character,character}.

See signature \code{character,character}.

See signature \code{character,character}.
}
\description{
Get package description from \code{DESCRIPTION} file.
}
\section{Methods (by class)}{
\itemize{
\item \code{from = character,fields = character}: 

\item \code{from = missing,fields = character}: 

\item \code{from = missing,fields = missing}: 

\item \code{from = character,fields = missing}: 
}}
\examples{

}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp.core.repos}
}

