% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getPackageTitle}
\alias{getPackageTitle}
\alias{getPackageTitle,character-method}
\alias{getPackageTitle,missing-method}
\title{Get Package Title}
\usage{
getPackageTitle(from = ".", ...)

\S4method{getPackageTitle}{character}(from = ".", ...)

\S4method{getPackageTitle}{missing}(from = ".", ...)
}
\arguments{
\item{from}{\strong{Signature argument}.
Object containing source information.}

\item{...}{Further arguments passed to subsequente functions:
\code{\link[rapp.core.description]{getPackageDescription}}.}

\item{from}{\code{\link{character}}.}

\item{from}{\code{\link{missing}}.}
}
\value{
\code{character}.

See signature \code{character}.
}
\description{
Get package title.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: 

\item \code{missing}: 
}}
\examples{

}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp.core.description}
}

